requirements.txt
cryptography
pyperclip
qrcode
pillow